+++ nis/nis_intern.h
@@ -41,18 +41,18 @@ extern long __nis_findfastest __P ((dir_binding *bind));
 extern nis_error __do_niscall2 __P ((const nis_server *serv, u_int serv_len,
 				     u_long prog, xdrproc_t xargs, caddr_t req,
 				     xdrproc_t xres, caddr_t resp,
-				     u_long flags, nis_cb *cb));
+				     unsigned int flags, nis_cb *cb));
 extern nis_error __do_niscall __P ((const_nis_name name, u_long prog,
 				    xdrproc_t xargs, caddr_t req,
 				    xdrproc_t xres, caddr_t resp,
-				    u_long flags, nis_cb *cb));
+				    unsigned int flags, nis_cb *cb));
 
 /* NIS+ callback */
 extern nis_error __nis_do_callback __P ((struct dir_binding *bptr,
 					 netobj *cookie, struct nis_cb *cb));
 extern struct nis_cb *__nis_create_callback
       __P ((int (*callback)(const_nis_name, const nis_object *, const void *),
-	    const void *userdata, u_long flags));
+	    const void *userdata, unsigned int flags));
 extern nis_error __nis_destroy_callback __P ((struct nis_cb *cb));
 
 __END_DECLS
