+++ nis/rpcsvc/ypupd.h
@@ -69,19 +69,19 @@ typedef struct ypdelete_args ypdelete_args;
 
 extern  bool_t xdr_ypdelete_args __P ((XDR *, ypdelete_args*));
 
-#define YPU_PROG ((u_long)100028)
-#define YPU_VERS ((u_long)1)
+#define YPU_PROG 100028
+#define YPU_VERS 1
 
-#define YPU_CHANGE ((u_long)1)
+#define YPU_CHANGE 1
 extern  u_int * ypu_change_1 __P ((ypupdate_args *, CLIENT *));
 extern  u_int * ypu_change_1_svc __P((ypupdate_args *, struct svc_req *));
-#define YPU_INSERT ((u_long)2)
+#define YPU_INSERT 2
 extern  u_int * ypu_insert_1 __P ((ypupdate_args *, CLIENT *));
 extern  u_int * ypu_insert_1_svc __P ((ypupdate_args *, struct svc_req *));
-#define YPU_DELETE ((u_long)3)
+#define YPU_DELETE 3
 extern  u_int * ypu_delete_1 __P ((ypdelete_args *, CLIENT *));
 extern  u_int * ypu_delete_1_svc __P ((ypdelete_args *, struct svc_req *));
-#define YPU_STORE ((u_long)4)
+#define YPU_STORE 4
 extern  u_int * ypu_store_1 __P ((ypupdate_args *, CLIENT *));
 extern  u_int * ypu_store_1_svc __P ((ypupdate_args *, struct svc_req *));
